requirements.txt
streamlit==1.8.1
torch==1.11.0
torchvision==0.12.0
Pillow==9.0.1
numpy==1.20.3
matplotlib==3.4.2
glob2==0.7
os
colabcode
